manifold/requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
Django
django-allauth
django-cors-headers
django-crontab
django-filter
django-multiselectfield
djangorestframework
flake8
idna
oauthlib
pandas
Pillow
pycparser
PyJWT
PyMySQL
python-barcode
python-decouple
python3-openid
pytz
qrcode
requests
requests-oauthlib
six
sqlparse
typing_extensions
tzdata
urllib
unicode 
escpos